Refactor various components for improved clarity and performance
- Simplified matrix multiplication methods by removing unnecessary cloning of elements.
- Optimized register writes in DISP3DCNT and other IO registers by directly using bitwise operations.
- Enhanced rendering logic by cleaning up debug print statements and improving variable handling.
- Streamlined VRAM access methods for better readability and efficiency.
- Updated interrupt controller to use direct bit manipulation for clarity.
- Refined IPC FIFO conditions to utilize more idiomatic Rust methods.
- Improved math operations by simplifying bitwise manipulations.
- Cleaned up memory read/write functions for ARM7 and ARM9 to enhance clarity.
- Adjusted GUI slider ranges to use more idiomatic Rust syntax.
- Added logging for savedata operations to improve debugging capabilities.

+++ core/src/hw/cartridge.rs
@@ -521,7 +521,7 @@ impl ROMCTRL {
             2 => {
                 (self.data_word_ready as u8) << 7
                     | (self.key2_encrypt_cmd as u8) << 6
-                    | (self.key1_gap2_len) as u8
+                    | (self.key1_gap2_len)
             }
             3 => {
                 (self.block_busy as u8) << 7
@@ -529,7 +529,7 @@ impl ROMCTRL {
                     | (self.resb_release_reset as u8) << 5
                     | (self.key1_gap_clks as u8) << 4
                     | (self.transfer_clk_rate as u8) << 3
-                    | (self.data_block_size as u8)
+                    | self.data_block_size
             }
             _ => unreachable!(),
         }
